feat: 16-bit float pipeline + PSD smart-object/PSB import
Internal pixel storage moves to linear RGBA32F end-to-end (was RGBA8):
- PixelLayer / RasterTiles / StrokeSession / brush + all tools /
  compositor upload + readback now float; .sable stores raw at the
  document depth (8 -> byte, 16 -> ushort, 32 -> float)
- import preserves 16-bit: self-contained 16-bit PNG encoder/decoder
  (Skia strips 16-bit) + 16-bit TIFF; PSD overbright bug fixed
- export writes true 16-bit PNG/TIFF at the document depth
- fixes latent byte->float overbright bugs (CreateDemo, OpenImage,
  PsdReader, GPU brush readback, Gpu.Spike)

PSD compatibility (P2 closers):
- Smart Object Tier 1: parse SoLd/SoLE descriptor -> capture placement
  transform + identity + source size into Layer.SmartObject (was
  discarded with a bare note)
- PSB Tier 1: accept version 2 + 64-bit section/channel lengths, raise
  the dim cap, gated on an 80 MP in-memory budget (honest rejection
  above it; full size needs tiled CPU storage)

573 tests pass; GPU smokes green.

+++ src/Sable.Engine/Commands/CropCommand.cs
@@ -14,7 +14,7 @@ public sealed class CropCommand : IUndoableCommand
     private readonly Document _doc;
     private readonly int _x, _y, _w, _h;
     private int _oldW, _oldH;
-    private readonly List<(PixelLayer layer, byte[] px, int w, int h, int offX, int offY)> _pixSnap = new();
+    private readonly List<(PixelLayer layer, float[] px, int w, int h, int offX, int offY)> _pixSnap = new();
     private readonly List<(Layer layer, byte[]? mask)> _maskSnap = new();
     private bool _captured;
